Simplify function for creating docs from React classes
